Integration of repo list downloader

+++ src/python/get_repos.py
@@ -8,11 +8,15 @@
 import gzip
 import urllib.request
 from io import BytesIO
+import os
 
 import pandas as pd
 import numpy as np
 
 if __name__ == "__main__":
+    if os.path.isfile("data/repos.csv"):
+        exit(0)
+
     urllib.request.urlretrieve(
         "https://reporeapers.github.io/static/downloads/dataset.csv.gz",
         "data/repos.csv.gz",
